refactor: add StatusCode and fix logRequest
