+++ src/utils/handler.ts
@@ -1,10 +1,14 @@
 import * as http from 'http';
 
-import { baseUrl, errorMessages, headers } from '../common/config.js';
 import { usersController } from '../controller/users.controller.js';
-import { isValidId } from './validateId.js';
-import { isValidUser } from './validateUser.js';
+import { isValidId, isValidUser } from '../middlewares/index.js';
 import { response } from './response.js';
+import {
+  baseUrl,
+  headers,
+  ValidateMessage,
+  StatusCode,
+} from '../common/config.js';
 
 export class Handler {
   baseUrl = baseUrl;
@@ -14,8 +18,8 @@ export class Handler {
     const { method, url } = req;
 
     if (!url?.startsWith(this.baseUrl)) {
-      response(req, res, 404, this.headers, {
-        message: errorMessages.endpoint,
+      response(req, res, StatusCode.NOT_FOUND, this.headers, {
+        message: ValidateMessage.ENDPOINT,
       });
       return;
     }
@@ -27,8 +31,8 @@ export class Handler {
         const id = url?.split('/').pop();
 
         if (!id || !isValidId(id)) {
-          response(req, res, 400, this.headers, {
-            message: errorMessages.uuid,
+          response(req, res, StatusCode.BAD_REQUEST, this.headers, {
+            message: ValidateMessage.UUID,
           });
         } else usersController.getUser(req, res, id).then(() => {});
       }
@@ -36,15 +40,15 @@ export class Handler {
 
     if (method === 'POST') {
       if (url !== this.baseUrl) {
-        response(req, res, 404, this.headers, {
-          message: errorMessages.endpoint,
+        response(req, res, StatusCode.NOT_FOUND, this.headers, {
+          message: ValidateMessage.UUID,
         });
       }
       let data = '';
 
       req.on('data', (chunk) => (data += chunk));
       req.on('error', (err) => {
-        response(req, res, 500, this.headers, {
+        response(req, res, StatusCode.INTERNAL_SERVER_ERROR, this.headers, {
           message: `Error has been occurs. ${err.message}`,
         });
       });
@@ -52,8 +56,8 @@ export class Handler {
         const user = JSON.parse(data);
 
         if (!isValidUser(user)) {
-          response(req, res, 400, this.headers, {
-            message: errorMessages.body,
+          response(req, res, StatusCode.BAD_REQUEST, this.headers, {
+            message: ValidateMessage.BODY,
           });
         } else usersController.addUser(req, res, user).then(() => {});
       });
@@ -63,14 +67,17 @@ export class Handler {
       const id = url?.split('/').pop();
 
       if (!id || !isValidId(id)) {
-        response(req, res, 400, this.headers, { message: errorMessages.uuid });
+        response(req, res, StatusCode.BAD_REQUEST, this.headers, {
+          message: ValidateMessage.UUID,
+        });
+        return;
       }
 
       let data = '';
 
       req.on('data', (chunk) => (data += chunk));
       req.on('error', (err) => {
-        response(req, res, 500, this.headers, {
+        response(req, res, StatusCode.INTERNAL_SERVER_ERROR, this.headers, {
           message: `Error has been occurs. ${err.message}`,
         });
       });
@@ -78,8 +85,8 @@ export class Handler {
         const user = JSON.parse(data);
 
         if (!isValidUser(user)) {
-          response(req, res, 400, this.headers, {
-            message: errorMessages.body,
+          response(req, res, StatusCode.BAD_REQUEST, this.headers, {
+            message: ValidateMessage.BODY,
           });
         } else {
           usersController.updateUser(req, res, { ...user, id }).then(() => {});
@@ -91,7 +98,9 @@ export class Handler {
       const id = url?.split('/').pop();
 
       if (!id || !isValidId(id)) {
-        response(req, res, 400, this.headers, { message: errorMessages.uuid });
+        response(req, res, StatusCode.BAD_REQUEST, this.headers, {
+          message: ValidateMessage.UUID,
+        });
       } else {
         usersController.deleteUser(req, res, id).then(() => {});
       }
